CONFIG: update type casting for pointer conversions across multiple files

+++ kernel/futex.c
@@ -3006,7 +3006,7 @@ static long futex_wait_restart(struct restart_block *restart)
 				restart->futex.val, tp, restart->futex.bitset);
 #else /* OPLUS_FEATURE_UIFIRST */
 	return (long)futex_wait(uaddr, restart->futex.flags,
-				restart->futex.val, tp, restart->futex.bitset, (u32)(restart->futex.uaddr2));
+				restart->futex.val, tp, restart->futex.bitset, (u32)(uintptr_t)(restart->futex.uaddr2));
 #endif /* OPLUS_FEATURE_UIFIRST */
 }
 
@@ -4010,7 +4010,7 @@ long do_futex(u32 __user *uaddr, int op, u32 val, ktime_t *timeout,
 // XieLiujie@BSP.KERNEL.PERFORMANCE, 2020/05/25, Add for UIFirst
 	return futex_wait(uaddr, flags, val, timeout, val3);
 #else /* OPLUS_FEATURE_UIFIRST */
-	return futex_wait(uaddr, flags, val, timeout, val3, (u32)uaddr2);
+	return futex_wait(uaddr, flags, val, timeout, val3, (u32)(uintptr_t)uaddr2);
 #endif /* OPLUS_FEATURE_UIFIRST */
 	case FUTEX_WAKE:
 		val3 = FUTEX_BITSET_MATCH_ANY;
